% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{ggWebGLOutput}
\alias{ggWebGLOutput}
\title{Shiny Output Binding for ggWebGL}
\usage{
ggWebGLOutput(outputId, width = "100\%", height = "480px")
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Widget dimensions.}
}
\value{
A Shiny output tag.
}
\description{
Shiny Output Binding for ggWebGL
}
\examples{
\dontshow{if (requireNamespace("shiny", quietly = TRUE)) withAutoprint(\{ # examplesIf}
ui <- shiny::fluidPage(
  ggWebGLOutput("plot", height = "320px")
)

ui
\dontshow{\}) # examplesIf}
}
