% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{renderGgWebGL}
\alias{renderGgWebGL}
\title{Render a ggWebGL Widget in Shiny}
\usage{
renderGgWebGL(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression returning a \code{ggWebGL} widget.}

\item{env}{Evaluation environment.}

\item{quoted}{Whether \code{expr} is quoted.}
}
\value{
A Shiny render function.
}
\description{
Render a ggWebGL Widget in Shiny
}
\examples{
\dontshow{if (requireNamespace("shiny", quietly = TRUE)) withAutoprint(\{ # examplesIf}
server <- function(input, output, session) {
  output$plot <- renderGgWebGL({
    ggplot_webgl(
      ggplot2::ggplot(
        mtcars[1:8, ],
        ggplot2::aes(mpg, wt, colour = factor(cyl))
      ) +
        geom_point_webgl(size = 2) +
        theme_webgl()
    )
  })
}

server
\dontshow{\}) # examplesIf}
}
